requirements.txt
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
ipykernel==6.23.3
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
matplotlib-inline==0.1.6
mysql-connector-python==8.0.33
nest-asyncio==1.5.6
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.8.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.1.0
six==1.16.0
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
wcwidth==0.2.6
